+++ app/http/Handler.js
@@ -1,8 +1,11 @@
 import config from "../../config/app.js";
 import loggerMiddleware from "./middleware/logger.js";
 import responseMacroMiddleware from "./middleware/response-macro.js";
+import connectFlash from "connect-flash";
 import cookieParser from "cookie-parser";
 import express from "express";
+import expressSession from "express-session";
+import methodOverride from "method-override";
 
 export class Handler {
   /**
@@ -23,7 +26,27 @@ export class Handler {
         limit: config.maximum_request_body_size,
       }),
     );
-    app.use(cookieParser());
+    app.use(cookieParser(config.cookie_key));
+    app.use(
+      expressSession({
+        secret: config.session_key,
+        cookie: { maxAge: config.cookie_max_age },
+        saveUninitialized: true,
+        resave: true,
+      }),
+    );
+    app.use(connectFlash());
+    app.use(
+      methodOverride(function (req) {
+        if (req.body && typeof req.body === "object" && "_method" in req.body) {
+          const method = req.body._method;
+
+          delete req.body._method;
+
+          return method;
+        }
+      }),
+    );
   }
 
   /**